Rename filter.js to regular-filter-modal.js
