+++ assets/js/dashboard/stats/modals/regular-filter-modal.js
